Додано домашнє завдання 14: Реєстрація - тест на валідацію порожнніх полів логіна, пошти, пароля - тест на валідацію існуючого логіна - тест на довжину логіна Вхід: - тест на коректний логін, пароль - вхід - тест на підтримку вводу логіна, пароля клавішею Enter - розділив тест на пусті логін, пароль на два тести - розділив тест на некоректні логін, пароль на два тести
