--- homework_makarovych_L10.txt
@@ -1,35 +1,0 @@
-"""Lesson 10 Homework"""
-
-Портал, що тестуємо - https://qa-complex-app-for-testing.herokuapp.com/
-
-- Написати 5 тест кейсів для стартової сторінки (не автоматизованих, а звичайних, які ми потім звісно автоматизуємо 🙂).
-Тільки давайте без "Перевітири, що є кнопка логін"  і тд. Щось функціональне та умовно корисне ). Формат будь-який, звичний вам.
-- Виписати до кожного тесту 1-2 локатори які для нього знадобляться (можна і більше 🙂)
-
-Тесткейси
-1. h1. Перевірити текст, шрифт, розмір шрифта, колір
-$x(".//h1[@class='display-3']")
-
-2. Форма реєстрація. Валідація на порожні поля логіна, почти, пароля
-$x(".//button[@class='py-3 mt-4 btn btn-lg btn-success btn-block']")
-$x(".//div[contains(text(),'Username must be at least 3 characters.')]")
-$x(".//div[contains(text(),'You must provide a valid email address.')]")
-$x(".//div[contains(text(),'Password must be at least 12 characters.')]")
-
-3. Форма реєстрації. Перевірка наявності підказок в полях
-$x(".//input[@id='username-register']")
-$x(".//input[@id='email-register']")
-$x(".//input[@id='password-register']")
-
-4. Форма входу. Валідація на введення некоректних логіна та(або) пароля
-$x(".//input[@id='username-register']")
-$x(".//input[@id='email-register']")
-
-$x(".//input[@class='form-control form-control-sm input-dark' and @name='password']")
-$x(".//button[@class='btn btn-primary btn-sm']")
-$x(".//div[@class='alert alert-danger text-center']")
-
-5. Форма входу. Кнопка Singn In переводить в аккаунт
-$x(".//button[@class='btn btn-primary btn-sm']")
-$x(".//button[@class='btn btn-sm btn-secondary']")
-$x(".//a[@href='/create-post']")
